chore: move meta tag to my-ads route
