+++ packages/p2p/src/components/buy-sell/buy-sell.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { observer } from 'mobx-react-lite';
-import { changeMetaTagWithOG } from '@deriv/shared';
 import { localize } from 'Components/i18next';
 import PageReturn from 'Components/page-return/page-return.jsx';
 import Verification from 'Components/verification/verification.jsx';
@@ -20,19 +19,7 @@ const BuySell = () => {
         const disposeAdvertIntervalReaction = buy_sell_store.registerAdvertIntervalReaction();
         buy_sell_store.setLocalCurrency(buy_sell_store.selected_local_currency);
 
-        const description_content = 'P2P Description';
-        const title_content = 'Deriv P2P';
-        const image =
-            'https://play-lh.googleusercontent.com/ah8RkaAnph2gouJ48fVeybeJgw-tu2dzTDYL7miccIWxvd0ZcK5-MM20bGxjpjb2lXU';
-
-        const restoreMetaTagWithOGDescription = changeMetaTagWithOG('description', description_content);
-        const restoreMetaTagWithOGTitle = changeMetaTagWithOG('title', title_content);
-        const restoreMetaTagWithOGImage = changeMetaTagWithOG('image', image);
-
         return () => {
-            restoreMetaTagWithOGDescription();
-            restoreMetaTagWithOGTitle();
-            restoreMetaTagWithOGImage();
             disposeIsListedReaction();
             disposeAdvertIntervalReaction();
         };
